Refactor: Inline `current_unser` into `ApplicationController`
The knowledge of whomever is the current user, is more of a controller
responsibility than a helper one. Since the controllers performs
authorizations that the current user is allowed to perform an action,
such as updating a user. The views can access the helpers in the
controllers. if they are exported with `helper_method`.