Add `incorrect_password?` to verify new password
We are using `has_secure_password`, a rails feature, to handle password
for a user. However we also configed `allow_nil: true` to make it not
necessary to fill in the password when you update your profile.

Because of this, the app will not complain nor update password_digest
if you reset your password with "", less than 6 characters or password
does't match.

To solve this issue, we added a `incorrect_password?` method to check
the new password in `password_reset_controller` and returns an error,
but the error will not specify why the password is invalid because
that will need a if elseif statement which is not very nice.

Below is the error message and it should be sufficient:
'Invalid password, minimum 6 characters and password_confirmation must
match'