WIP: Refactor `ResetPasswordController`
Want to refactor `validate_user` and `check_expiration` methods to a
service object. After refactor, the `redirect_to` `login_path` `flash`
does not work any more. I tried to include the files where those
methods are wrote. but it is give further errors which I think I am
taking the wrong direction to do this. This is a WIP commit.

error message:
undefined local variable or method `login_path' for
#<ValidateRequest:0x00007f884cfed108>