Add `forget password` feature (#14)
**Background:**
User might forget their password and want to reset it. We do it by 
setting a `password_reset_digest` attribute for user and send an email 
to user registered email address with a link to password reset page. 
When user clicks the link, we find the user by the email included in the 
url and authenticate by comparing the token in the url and the user's
`password_reset_digest`.
If the new password is valid, we update it and redirects to login page.
If not, show an error message. If the `password_reset_digest` is
generated more than 30 minutes, redirect to `new_edit_password_path`

In development, using `MailCatcher` to catch password reset emails.

**Ticket on Monday.com**
https://universalavenue.monday.com/boards/173844071/pulses/173851243

**Changes:**
- Add `PasswordResetController` to verify and update password
- Add `UserMailer` to send password reset email
- Add `Token` class to generate token
- Add `Digest` class to convert a token to digest
- Refactor render error messages to `application.html.erb`
- Rename flash[:danger] to flash[:error]
- Update default `from` email address for mailer
- Add `ActiveSupport::Testing::TimeHelpers` in `rails_helper.rb`
- Config MailCatcher server
- Add `forget_password` and `reset_password` in `AuthHelper.rb`