Add `validate_time_format` method
Add the method in `ValidateEvent` -> `TimeValidator`.
Add a regex to match the given `start_time` `end_time` format.
Add `validate_time_format` before_action in `EventController`
to check the format.
Call the method from `EventController` and gives an error if
one of the format is not match.
Refactor `ValidateEvenet.new` to be called in `validate_timeformat`.
Fix "formate" typo, without "e".