User creates event (#15)
**Background:**
User should be able to create an event with "Title" "Start at" 
"End at" "Location" and "Description". We added a button 
"Create Event" on home page, it shows the create form if user
is logged in, otherwise, redirects to login page. Once an event
is created, renders the "show" page.
Next step is to set up the "index" page so we can redirects
to "index" page instead of "show" page.

**Tickets on Monday.com**
https://universalavenue.monday.com/boards/173844071/pulses/173851742

**Changes**
- Add `Event` model/controller/views
- Refactor `logged_in_user` from `UserController` to 
  `ApplicationController`
- Add `InFutureValidator` to validate `start_at` and `end_at`
  to be a future date
- Add a method in `Event` model to validate `end_at` should
  not be earlier than `start_at`.
- Use ActiveSupport core extensions to get i.e `1.hour.from_now`.