Add search feature using pg_search_scope (#18)
### Background:
We want user to be able to search for an event or a user. We added 
a search bar in events index page, it can take one or multiple
words and search through both Events and Users table. Using 
"tsearch" and "trigram" search algorithm.

The "title" "locaiton" and "description" columns in Event table and 
"name"  "email" in User table are searchable. 

### Ticket on Monday.com
https://universalavenue.monday.com/boards/173844071/pulses/173852473

### Changes
- Add pg_search_scope `search` for both Event and User
- Add index for "title" "location" "description" columns in Event
- Add index for "name" in User
- Add SearchesController and Search service to 
   handle search request
- Add search bar in event index view
- Update Rubocop to ignore multiple expectations and nesting